compressor/zstd: add zstd to embedded ceph
